Fix the delete layanan feature
