+++ resources/views/layouts/merchant.blade.php
@@ -171,8 +171,6 @@
             {{ $slot }}
         </main>
     </div>
-
-    @stack('modals')
     @livewireScripts
 </body>
 </html>
