+++ resources/views/merchant/services.blade.php
@@ -33,108 +33,7 @@
                 </div>
 
                 @if (count($layanan) > 0)
-                    <div class="overflow-x-auto">
-                        <table class="min-w-full divide-y divide-gray-200">
-                            <thead class="bg-gray-50">
-                                <tr>
-                                    <th
-                                        class="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                                        Nama Layanan</th>
-                                    <th
-                                        class="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                                        Harga</th>
-                                    <th
-                                        class="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                                        Durasi</th>
-                                    <th
-                                        class="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                                        Jam Operasional</th>
-                                    <th
-                                        class="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                                        Status</th>
-                                    <th
-                                        class="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                                        Aksi</th>
-                                </tr>
-                            </thead>
-                            <tbody class="bg-white divide-y divide-gray-200">
-                                @foreach ($layanan as $item)
-                                    <tr>
-                                        <td class="px-6 py-4 whitespace-nowrap">
-                                            <div class="flex items-center">
-                                                <div class="flex-shrink-0 h-10 w-10">
-                                                    @if ($item->aset)
-                                                        <img class="h-10 w-10 rounded-full object-contain"
-                                                            src="{{ $item->aset->media_url }}"
-                                                            alt="{{ $item->nama_layanan }}">
-                                                    @else
-                                                        <div
-                                                            class="h-10 w-10 rounded-full bg-homize-gray flex items-center justify-center">
-                                                            <svg class="h-6 w-6 text-gray-400" fill="none"
-                                                                stroke="currentColor" viewBox="0 0 24 24">
-                                                                <path stroke-linecap="round" stroke-linejoin="round"
-                                                                    stroke-width="2"
-                                                                    d="M4 16l4.586-4.586a2 2 0 012.828 0L16 16m-2-2l1.586-1.586a2 2 0 012.828 0L20 14m-6-6h.01M6 20h12a2 2 0 002-2V6a2 2 0 00-2-2H6a2 2 0 00-2 2v12a2 2 0 002 2z">
-                                                                </path>
-                                                            </svg>
-                                                        </div>
-                                                    @endif
-                                                </div>
-                                                <div class="ml-4">
-                                                    <div class="text-sm font-medium text-gray-900">
-                                                        {{ $item->nama_layanan }}</div>
-                                                    <div class="text-sm text-gray-500">
-                                                        {{ Str::limit($item->deskripsi_layanan, 50) }}</div>
-                                                </div>
-                                            </div>
-                                        </td>
-                                        <td class="px-6 py-4 whitespace-nowrap">
-                                            @if ($item->tarif_layanan)
-                                                <div class="text-sm text-gray-900">Rp
-                                                    {{ number_format($item->tarif_layanan->harga, 0, ',', '.') }}</div>
-                                                <div class="text-sm text-gray-500">{{ $item->tarif_layanan->satuan }}
-                                                </div>
-                                            @else
-                                                <div class="text-sm text-gray-500">Belum ada tarif</div>
-                                            @endif
-                                        </td>
-                                        <td class="px-6 py-4 whitespace-nowrap">
-                                            @if ($item->tarif_layanan)
-                                                <div class="text-sm text-gray-900">{{ $item->tarif_layanan->durasi }}
-                                                    {{ $item->tarif_layanan->tipe_durasi }}</div>
-                                            @else
-                                                <div class="text-sm text-gray-500">-</div>
-                                            @endif
-                                        </td>
-                                        <td class="px-6 py-4 whitespace-nowrap">
-                                            <div class="text-sm text-gray-900">
-                                                {{ \App\Helpers\HariHelper::formatHari($item->jam_operasional->hari->pluck('nama_hari')->implode(',')) }}
-                                            </div>
-                                            <div class="text-sm text-gray-500">
-                                                {{ $item->jam_operasional->jam_buka }} -
-                                                {{ $item->jam_operasional->jam_tutup }}
-                                            </div>
-                                        </td>
-                                        <td class="px-6 py-4 whitespace-nowrap">
-                                            <span
-                                                class="px-2 inline-flex text-xs leading-5 font-semibold rounded-full bg-green-100 text-green-800">
-                                                Aktif
-                                            </span>
-                                        </td>
-                                        <td class="px-6 py-4 whitespace-nowrap text-sm font-medium">
-                                            <a href="{{ route('layanan.detail', $item->id) }}"
-                                                class="text-homize-blue hover:text-homize-blue-second mr-3">Lihat</a>
-                                            <a href="#"
-                                                class="text-indigo-600 hover:text-indigo-900 mr-3 edit-layanan"
-                                                data-id="{{ $item->id }}">Edit</a>
-                                            <a href="#" class="text-red-600 hover:text-red-900 delete-layanan"
-                                                data-id="{{ $item->id }}">Hapus</a>
-                                        </td>
-                                    </tr>
-                                @endforeach
-                            </tbody>
-                        </table>
-                    </div>
+                    @include('merchant.components.tabel-layanan')
                 @else
                     <div class="text-center py-12">
                         <div class="mb-4">
@@ -191,227 +90,7 @@
             </div>
 
             <!-- Form edit layanan -->
-            <form id="editLayananForm" method="POST" enctype="multipart/form-data">
-                @csrf
-                @method('PUT')
-                <div class="space-y-4">
-                    <!-- Basic Information -->
-                    <div>
-                        <h4 class="text-md font-medium text-gray-700 mb-2">Informasi Dasar</h4>
-                        <div class="grid grid-cols-1 gap-4">
-                            <div>
-                                <label for="edit_nama_layanan" class="block text-sm font-medium text-gray-700">Nama
-                                    Layanan</label>
-                                <input type="text" name="nama_layanan" id="edit_nama_layanan"
-                                    class="mt-1 focus:ring-homize-blue focus:border-homize-blue block w-full shadow-sm sm:text-sm border-gray-300 rounded-md"
-                                    required>
-                            </div>
-                            <div>
-                                <label for="edit_deskripsi_layanan"
-                                    class="block text-sm font-medium text-gray-700">Deskripsi Layanan</label>
-                                <textarea name="deskripsi_layanan" id="edit_deskripsi_layanan" rows="3"
-                                    class="mt-1 focus:ring-homize-blue focus:border-homize-blue block w-full shadow-sm sm:text-sm border-gray-300 rounded-md"
-                                    required></textarea>
-                            </div>
-                            <div>
-                                <label for="edit_id_sub_kategori" class="block text-sm font-medium text-gray-700">Sub
-                                    Kategori</label>
-                                <select name="id_sub_kategori" id="edit_id_sub_kategori"
-                                    class="mt-1 block w-full py-2 px-3 border border-gray-300 bg-white rounded-md shadow-sm focus:outline-none focus:ring-homize-blue focus:border-homize-blue sm:text-sm"
-                                    required>
-                                    @foreach ($subKategori as $sub)
-                                        <option value="{{ $sub->id }}">{{ $sub->nama }}</option>
-                                    @endforeach
-                                </select>
-                            </div>
-                            <div>
-                                <label for="edit_pengalaman"
-                                    class="block text-sm font-medium text-gray-700">Pengalaman (Tahun)</label>
-                                <input type="number" name="pengalaman" id="edit_pengalaman" min="0"
-                                    class="mt-1 focus:ring-homize-blue focus:border-homize-blue block w-full shadow-sm sm:text-sm border-gray-300 rounded-md">
-                            </div>
-                        </div>
-                    </div>
-
-                    <!-- Pricing -->
-                    <div>
-                        <h4 class="text-md font-medium text-gray-700 mb-2">Harga & Durasi</h4>
-                        <div class="grid grid-cols-1 md:grid-cols-2 gap-4">
-                            <div>
-                                <label for="edit_harga" class="block text-sm font-medium text-gray-700">Harga</label>
-                                <div class="mt-1 relative rounded-md shadow-sm">
-                                    <div class="absolute inset-y-0 left-0 pl-3 flex items-center pointer-events-none">
-                                        <span class="text-gray-500 sm:text-sm">Rp</span>
-                                    </div>
-                                    <input type="number" name="harga" id="edit_harga"
-                                        class="focus:ring-homize-blue focus:border-homize-blue block w-full pl-10 pr-12 sm:text-sm border-gray-300 rounded-md"
-                                        placeholder="0" required>
-                                </div>
-                            </div>
-                            <div>
-                                <label for="edit_satuan"
-                                    class="block text-sm font-medium text-gray-700">Satuan</label>
-                                <select name="satuan" id="edit_satuan"
-                                    class="mt-1 block w-full py-2 px-3 border border-gray-300 bg-white rounded-md shadow-sm focus:outline-none focus:ring-homize-blue focus:border-homize-blue sm:text-sm">
-                                    <option value="kg"
-                                        {{ isset($editLayanan) && $editLayanan->tarif_layanan && $editLayanan->tarif_layanan->satuan == 'kg' ? 'selected' : '' }}>
-                                        Per Kilogram (kg)</option>
-                                    <option value="unit"
-                                        {{ isset($editLayanan) && $editLayanan->tarif_layanan && $editLayanan->tarif_layanan->satuan == 'unit' ? 'selected' : '' }}>
-                                        Per Unit</option>
-                                    <option value="pcs"
-                                        {{ isset($editLayanan) && $editLayanan->tarif_layanan && $editLayanan->tarif_layanan->satuan == 'pcs' ? 'selected' : '' }}>
-                                        Per Pieces (pcs)</option>
-                                    <option value="jam"
-                                        {{ isset($editLayanan) && $editLayanan->tarif_layanan && $editLayanan->tarif_layanan->satuan == 'jam' ? 'selected' : '' }}>
-                                        Per Jam</option>
-                                </select>
-                                <div>
-                                    <label for="edit_durasi"
-                                        class="block text-sm font-medium text-gray-700">Durasi</label>
-                                    <input type="number" name="durasi" id="edit_durasi" min="1"
-                                        class="mt-1 block w-full py-2 px-3 border border-gray-300 bg-white rounded-md shadow-sm focus:outline-none focus:ring-homize-blue focus:border-homize-blue sm:text-sm"
-                                        required>
-                                </div>
-                                <div>
-                                    <label for="edit_tipe_durasi" class="block text-sm font-medium text-gray-700">Tipe
-                                        Durasi</label>
-                                    <select name="tipe_durasi" id="edit_tipe_durasi"
-                                        class="mt-1 block w-full py-2 px-3 border border-gray-300 bg-white rounded-md shadow-sm focus:outline-none focus:ring-homize-blue focus:border-homize-blue sm:text-sm"
-                                        required>
-                                        <option value="Jam"
-                                            {{ isset($editLayanan) && $editLayanan->tarif_layanan && $editLayanan->tarif_layanan->tipe_durasi == 'Jam' ? 'selected' : '' }}>
-                                            Jam</option>
-                                        <option value="Hari"
-                                            {{ isset($editLayanan) && $editLayanan->tarif_layanan && $editLayanan->tarif_layanan->tipe_durasi == 'Hari' ? 'selected' : '' }}>
-                                            Hari</option>
-                                        <option value="Pertemuan"
-                                            {{ isset($editLayanan) && $editLayanan->tarif_layanan && $editLayanan->tarif_layanan->tipe_durasi == 'Pertemuan' ? 'selected' : '' }}>
-                                            Pertemuan</option>
-                                    </select>
-                                </div>
-                            </div>
-                        </div>
-
-                        <!-- Jam Operasional -->
-                        <div>
-                            <h4 class="text-md font-medium text-gray-700 mb-2">Jam Operasional</h4>
-                            <div class="grid grid-cols-1 md:grid-cols-2 gap-4 mb-4">
-                                <div>
-                                    <label for="edit_jam_buka" class="block text-sm font-medium text-gray-700">Jam
-                                        Buka</label>
-                                    <input type="time" name="jam_operasional[jam_buka]" id="edit_jam_buka"
-                                        class="mt-1 focus:ring-homize-blue focus:border-homize-blue block w-full shadow-sm sm:text-sm border-gray-300 rounded-md"
-                                        required>
-                                </div>
-                                <div>
-                                    <label for="edit_jam_tutup" class="block text-sm font-medium text-gray-700">Jam
-                                        Tutup</label>
-                                    <input type="time" name="jam_operasional[jam_tutup]" id="edit_jam_tutup"
-                                        class="mt-1 focus:ring-homize-blue focus:border-homize-blue block w-full shadow-sm sm:text-sm border-gray-300 rounded-md"
-                                        required>
-                                </div>
-                            </div>
-                            <div>
-                                <label class="block text-sm font-medium text-gray-700 mb-2">Hari Operasional</label>
-                                <div class="grid grid-cols-2 md:grid-cols-4 gap-2">
-                                    <div class="flex items-center">
-                                        <input type="checkbox" name="jam_operasional[hari][]" value="1"
-                                            id="edit_hari_senin"
-                                            class="h-4 w-4 text-homize-blue focus:ring-homize-blue border-gray-300 rounded">
-                                        <label for="edit_hari_senin"
-                                            class="ml-2 block text-sm text-gray-700">Senin</label>
-                                    </div>
-                                    <div class="flex items-center">
-                                        <input type="checkbox" name="jam_operasional[hari][]" value="2"
-                                            id="edit_hari_selasa"
-                                            class="h-4 w-4 text-homize-blue focus:ring-homize-blue border-gray-300 rounded">
-                                        <label for="edit_hari_selasa"
-                                            class="ml-2 block text-sm text-gray-700">Selasa</label>
-                                    </div>
-                                    <div class="flex items-center">
-                                        <input type="checkbox" name="jam_operasional[hari][]" value="3"
-                                            id="edit_hari_rabu"
-                                            class="h-4 w-4 text-homize-blue focus:ring-homize-blue border-gray-300 rounded">
-                                        <label for="edit_hari_rabu"
-                                            class="ml-2 block text-sm text-gray-700">Rabu</label>
-                                    </div>
-                                    <div class="flex items-center">
-                                        <input type="checkbox" name="jam_operasional[hari][]" value="4"
-                                            id="edit_hari_kamis"
-                                            class="h-4 w-4 text-homize-blue focus:ring-homize-blue border-gray-300 rounded">
-                                        <label for="edit_hari_kamis"
-                                            class="ml-2 block text-sm text-gray-700">Kamis</label>
-                                    </div>
-                                    <div class="flex items-center">
-                                        <input type="checkbox" name="jam_operasional[hari][]" value="5"
-                                            id="edit_hari_jumat"
-                                            class="h-4 w-4 text-homize-blue focus:ring-homize-blue border-gray-300 rounded">
-                                        <label for="edit_hari_jumat"
-                                            class="ml-2 block text-sm text-gray-700">Jumat</label>
-                                    </div>
-                                    <div class="flex items-center">
-                                        <input type="checkbox" name="jam_operasional[hari][]" value="6"
-                                            id="edit_hari_sabtu"
-                                            class="h-4 w-4 text-homize-blue focus:ring-homize-blue border-gray-300 rounded">
-                                        <label for="edit_hari_sabtu"
-                                            class="ml-2 block text-sm text-gray-700">Sabtu</label>
-                                    </div>
-                                    <div class="flex items-center">
-                                        <input type="checkbox" name="jam_operasional[hari][]" value="7"
-                                            id="edit_hari_minggu"
-                                            class="h-4 w-4 text-homize-blue focus:ring-homize-blue border-gray-300 rounded">
-                                        <label for="edit_hari_minggu"
-                                            class="ml-2 block text-sm text-gray-700">Minggu</label>
-                                    </div>
-                                </div>
-                            </div>
-                        </div>
-
-                        <!-- Sertifikasi -->
-                        <div>
-                            <h4 class="text-md font-medium text-gray-700 mb-2">Sertifikasi</h4>
-                            <div class="grid grid-cols-1 gap-4">
-                                <div>
-                                    <label for="edit_nama_sertifikasi"
-                                        class="block text-sm font-medium text-gray-700">Nama Sertifikasi</label>
-                                    <input type="text" name="nama_sertifikasi" id="edit_nama_sertifikasi"
-                                        class="mt-1 focus:ring-homize-blue focus:border-homize-blue block w-full shadow-sm sm:text-sm border-gray-300 rounded-md">
-                                </div>
-                                <div>
-                                    <label for="edit_file_sertifikasi"
-                                        class="block text-sm font-medium text-gray-700">File Sertifikasi
-                                        (Opsional)</label>
-                                    <input type="file" name="file_sertifikasi" id="edit_file_sertifikasi"
-                                        class="mt-1 block w-full text-sm text-gray-500 file:mr-4 file:py-2 file:px-4 file:rounded-md file:border-0 file:text-sm file:font-semibold file:bg-homize-blue file:text-white hover:file:bg-homize-blue-second">
-                                    <p class="mt-1 text-sm text-gray-500" id="edit_current_sertifikasi"></p>
-                                </div>
-                            </div>
-                        </div>
-
-                        <!-- Gambar -->
-                        <div>
-                            <h4 class="text-md font-medium text-gray-700 mb-2">Gambar Layanan</h4>
-                            <div>
-                                <label for="edit_aset_layanan" class="block text-sm font-medium text-gray-700">Upload
-                                    Gambar (Opsional)</label>
-                                <input type="file" name="aset_layanan[]" id="edit_aset_layanan" multiple
-                                    class="mt-1 block w-full text-sm text-gray-500 file:mr-4 file:py-2 file:px-4 file:rounded-md file:border-0 file:text-sm file:font-semibold file:bg-homize-blue file:text-white hover:file:bg-homize-blue-second">
-                                <p class="mt-1 text-sm text-gray-500">Upload gambar baru akan menggantikan gambar yang
-                                    ada</p>
-                                <div id="edit_current_image" class="mt-2"></div>
-                            </div>
-                        </div>
-
-                        <div class="flex justify-end pt-4">
-                            <button type="button" id="cancelEditBtn"
-                                class="bg-white py-2 px-4 border border-gray-300 rounded-md shadow-sm text-sm font-medium text-gray-700 hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-homize-blue mr-2">Batal</button>
-                            <button type="submit"
-                                class="inline-flex justify-center py-2 px-4 border border-transparent shadow-sm text-sm font-medium rounded-md text-white bg-homize-blue hover:bg-homize-blue-second focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-homize-blue">Simpan
-                                Perubahan</button>
-                        </div>
-                    </div>
-            </form>
+            @include('merchant.components.edit-layanan')
         </div>
     </div>
 
@@ -441,212 +120,6 @@
         </div>
     </div>
 
-    <script>
-        document.addEventListener('DOMContentLoaded', function() {
-            const addModal = document.getElementById('layananModal');
-            const editModal = document.getElementById('editLayananModal');
-            const deleteModal = document.getElementById('deleteConfirmModal');
-            const addBtn = document.getElementById('addLayananBtn');
-            const addFirstBtn = document.getElementById('addFirstLayananBtn');
-            const closeAddBtn = document.getElementById('closeModal');
-            const closeEditBtn = document.getElementById('closeEditModal');
-            const cancelEditBtn = document.getElementById('cancelEditBtn');
-            const cancelDeleteBtn = document.getElementById('cancelDeleteBtn');
-            const confirmDeleteBtn = document.getElementById('confirmDeleteBtn');
-            const editLayananForm = document.getElementById('editLayananForm');
-
-            let currentLayananId = null;
-
-            function openAddModal() {
-                addModal.classList.remove('hidden');
-                document.body.style.overflow = 'hidden';
-            }
-
-            function closeAddModal() {
-                addModal.classList.add('hidden');
-                document.body.style.overflow = 'auto';
-            }
-
-            function openEditModal() {
-                editModal.classList.remove('hidden');
-                document.body.style.overflow = 'hidden';
-            }
-
-            function closeEditModal() {
-                editModal.classList.add('hidden');
-                document.body.style.overflow = 'auto';
-            }
-
-            function openDeleteModal() {
-                deleteModal.classList.remove('hidden');
-                document.body.style.overflow = 'hidden';
-            }
-
-            function closeDeleteModal() {
-                deleteModal.classList.add('hidden');
-                document.body.style.overflow = 'auto';
-            }
-
-            if (addBtn) addBtn.addEventListener('click', openAddModal);
-            if (addFirstBtn) addFirstBtn.addEventListener('click', openAddModal);
-            closeAddBtn.addEventListener('click', closeAddModal);
-            closeEditBtn.addEventListener('click', closeEditModal);
-            cancelEditBtn.addEventListener('click', closeEditModal);
-            cancelDeleteBtn.addEventListener('click', closeDeleteModal);
-
-            // Close modals when clicking outside
-            addModal.addEventListener('click', function(e) {
-                if (e.target === addModal) {
-                    closeAddModal();
-                }
-            });
-
-            editModal.addEventListener('click', function(e) {
-                if (e.target === editModal) {
-                    closeEditModal();
-                }
-            });
-
-            deleteModal.addEventListener('click', function(e) {
-                if (e.target === deleteModal) {
-                    closeDeleteModal();
-                }
-            });
-
-            // Add event listeners for edit buttons
-            document.querySelectorAll('.edit-layanan').forEach(button => {
-                button.addEventListener('click', function(e) {
-                    e.preventDefault();
-                    const layananId = this.getAttribute('data-id');
-                    currentLayananId = layananId;
-                    fetchLayananDetails(layananId);
-                });
-            });
-
-            // Add event listeners for delete buttons
-            document.querySelectorAll('.delete-layanan').forEach(button => {
-                button.addEventListener('click', function(e) {
-                    e.preventDefault();
-                    const layananId = this.getAttribute('data-id');
-                    currentLayananId = layananId;
-                    openDeleteModal();
-                });
-            });
-
-            // Fetch layanan details for editing
-            function fetchLayananDetails(id) {
-                fetch(`/merchant/layanan/${id}/edit`)
-                    .then(response => response.json())
-                    .then(data => {
-                        if (data.success) {
-                            populateEditForm(data);
-                            openEditModal();
-                        } else {
-                            alert('Error: ' + data.message);
-                        }
-                    })
-                    .catch(error => {
-                        console.error('Error:', error);
-                        alert('Terjadi kesalahan saat mengambil data layanan.');
-                    });
-            }
-
-            // Populate edit form with layanan data
-            function populateEditForm(data) {
-                const layanan = data.layanan;
-                const selectedHari = data.selectedHari;
-
-                // Set form action
-                editLayananForm.action = `/merchant/layanan/${layanan.id}`;
-
-                // Basic information
-                document.getElementById('edit_nama_layanan').value = layanan.nama_layanan;
-                document.getElementById('edit_deskripsi_layanan').value = layanan.deskripsi_layanan;
-                document.getElementById('edit_id_sub_kategori').value = layanan.id_sub_kategori;
-                document.getElementById('edit_pengalaman').value = layanan.pengalaman || 0;
-
-                // Pricing
-                if (layanan.tarif_layanan) {
-                    document.getElementById('edit_harga').value = layanan.tarif_layanan.harga;
-                    document.getElementById('edit_satuan').value = layanan.tarif_layanan.satuan;
-                    document.getElementById('edit_durasi').value = layanan.tarif_layanan.durasi;
-                    document.getElementById('edit_tipe_durasi').value = layanan.tarif_layanan.tipe_durasi;
-                }
-
-                // Jam operasional
-                if (layanan.jam_operasional) {
-                    document.getElementById('edit_jam_buka').value = layanan.jam_operasional.jam_buka;
-                    document.getElementById('edit_jam_tutup').value = layanan.jam_operasional.jam_tutup;
-
-                    // Reset all checkboxes first
-                    document.querySelectorAll('input[name="jam_operasional[hari][]"]').forEach(checkbox => {
-                        checkbox.checked = false;
-                    });
-
-                    // Check the appropriate days
-                    selectedHari.forEach(hariId => {
-                        const checkbox = document.querySelector(
-                            `input[name="jam_operasional[hari][]"][value="${hariId}"]`);
-                        if (checkbox) checkbox.checked = true;
-                    });
-                }
-
-                // Sertifikasi
-                if (layanan.sertifikasi && layanan.sertifikasi.length > 0) {
-                    document.getElementById('edit_nama_sertifikasi').value = layanan.sertifikasi[0]
-                        .nama_sertifikasi || '';
-                    if (layanan.sertifikasi[0].media_url) {
-                        document.getElementById('edit_current_sertifikasi').textContent = 'Sertifikasi saat ini: ' +
-                            layanan.sertifikasi[0].nama_sertifikasi;
-                    }
-                }
-
-                // Current image
-                const currentImageDiv = document.getElementById('edit_current_image');
-                currentImageDiv.innerHTML = '';
-                if (layanan.aset && layanan.aset.media_url) {
-                    const img = document.createElement('img');
-                    img.src = `${layanan.aset.media_url}`;
-                    img.alt = layanan.nama_layanan;
-                    img.className = 'h-24 w-24 object-cover rounded-md';
-                    currentImageDiv.appendChild(img);
-                }
-            }
-
-            // Handle delete confirmation
-            confirmDeleteBtn.addEventListener('click', function() {
-                if (currentLayananId) {
-                    deleteLayanan(currentLayananId);
-                }
-            });
-
-            // Delete layanan
-            function deleteLayanan(id) {
-                fetch(`/merchant/layanan/${id}`, {
-                        method: 'DELETE',
-                        headers: {
-                            'X-CSRF-TOKEN': document.querySelector('meta[name="csrf-token"]').getAttribute(
-                                'content'),
-                            'Content-Type': 'application/json'
-                        }
-                    })
-                    .then(response => response.json())
-                    .then(data => {
-                        closeDeleteModal();
-                        if (data.success) {
-                            // Show success message and reload page
-                            alert('Layanan berhasil dihapus');
-                            window.location.reload();
-                        } else {
-                            alert('Error: ' + data.message);
-                        }
-                    })
-                    .catch(error => {
-                        console.error('Error:', error);
-                        alert('Terjadi kesalahan saat menghapus layanan.');
-                        closeDeleteModal();
-                    });
-            }
-        });
-    </script>
+    @include('merchant.components.service-script')
+    
 </x-merchant-layout>
